refactor: #65 클래스 위치 이동 exception 패키지 내부로 들어갈 수 있도록 클래스 위치 수정
