+++ src/main/java/site/dogether/memberactivity/exception/InvalidParameterException.java
@@ -1,4 +1,4 @@
-package site.dogether.memberactivity;
+package site.dogether.memberactivity.exception;
 
 public class InvalidParameterException extends RuntimeException {
     public InvalidParameterException(final String message) {
